OLD_CODE_2/backend/requirements.txt
django-ninja

flower
gunicorn
docker
celery[pytest,redis]
cryptography
pycryptodome
redis
Jinja2

django-redis
python-dotenv
django-cors-headers
djangorestframework